requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.3.5
    # via aiohttp
aiohttp==3.10.3
    # via discord-py
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==24.2.0
    # via aiohttp
discord==2.3.2
    # via -r requirements.in
discord-py==2.4.0
    # via discord
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
gevent==24.2.1
    # via websocket
greenlet==3.0.3
    # via
    #   gevent
    #   websocket
idna==3.7
    # via yarl
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
pillow==10.4.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
websocket==0.2.1
    # via -r requirements.in
yarl==1.9.4
    # via aiohttp
zope-event==5.0
    # via gevent
zope-interface==7.0.1
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
